requirements.txt
# Core Dependencies
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5

# Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# ML & Forecasting
tensorflow==2.15.0
statsmodels==0.14.1
prophet==1.1.5

# Optimization
pulp==2.7.0
ortools==9.8.3296

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Visualization & Monitoring
plotly==5.18.0
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1

# Additional
python-dateutil==2.8.2
pytz==2023.3
jsonschema==4.20.0